+++ src/test/kotlin/dev/ekvedaras/laravelquery/completion/ColumnsCompletionTest.kt
@@ -207,4 +207,27 @@ internal class ColumnsCompletionTest : BaseTestCase() {
                 assertNoCompletion(*notExpected.toList().toTypedArray())
             }
     }
+
+    fun testDoesNotCompleteTableColumnsOtherTestCaseMethods() {
+        val table = DasUtil.getTables(dataSource()).first { it.name == "users" }
+        val columns = DasUtil.getColumns(table).map { it.name }
+        val lastTable = DasUtil.getTables(dataSource()).first { it.name == "customers" }
+
+        val notExpected =
+            schemas.filterNot { it == table.dasParent?.name } + // All other schemas
+                schemaTables.entries.filterNot { it.key == table.dasParent?.name }.map { it.value }
+                    .flatten() + // Tables of other schemas
+                DasUtil.getColumns(lastTable)
+                    .filterNot { columns.contains(it.name) }
+                    .map { it.name } // Columns of other table
+
+        myFixture.configureByText(
+            "test.php",
+            "<?php class Test extends \\Tests\\TestCase { public function test_it_completes() { \$this->get('users', ['<caret>']); } }"
+        )
+        myFixture.completeBasic()
+
+        assertNoCompletion(*columns.toList().toTypedArray())
+        assertNoCompletion(*notExpected.toList().toTypedArray())
+    }
 }